add import checkTotOH.py
